services/requirements.txt
# QUANTUM HUB BACKEND REQUIREMENTS
# =================================

# Core API Service Dependencies
# ---------------------------------
fastapi>=0.104.1
uvicorn>=0.24.0
motor>=3.3.1
pydantic>=2.5.0,<2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.1
python-dotenv>=1.0.0

# Quantum Backend Support
# ---------------------------------
# IBM Quantum
qiskit==1.4.1
qiskit-aer==0.16.4
qiskit-ibm-runtime==0.36.1
qiskit-ibm-provider==0.11.0

# AWS Braket
amazon-braket-sdk==1.90.1
amazon-braket-schemas==1.23.1
amazon-braket-default-simulator==1.26.4
boto3>=1.24.0  

# Google Quantum
cirq==1.4.1
cirq-core==1.4.1
cirq-google==1.4.1

# OpenQASM Support
openqasm3>=0.4.0
qiskit-qasm3-import==0.5.1
qiskit-braket-provider>=0.1.0

# Database & Storage
# ---------------------------------
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.5
redis>=5.0.0
minio>=7.2.0

# Monitoring & Logging
# ---------------------------------
prometheus-client>=0.19.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
structlog>=24.0.0
python-json-logger>=2.0.7

# Messaging & Streaming
# ---------------------------------
celery>=5.3.0             # Task queue
kafka-python>=2.0.0       # Event streaming
websockets>=12.0.0        # Real-time communication

# Scientific & Visualization
# ---------------------------------
numpy>=1.22.0
scipy>=1.8.0
pandas>=1.5.0
matplotlib>=3.5.0
networkx>=2.8.0
seaborn>=0.12.0

# Development & Testing
# ---------------------------------
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.10.1
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.1
tqdm>=4.65.0

# Security & Authentication
# ---------------------------------
cryptography>=40.0.0
requests>=2.28.0
pyjwt>=2.8.0
click>=8.1.0

# Note: Each service should have its own requirements.txt with just what it needs
# This master file is for reference and development environment setup
